chore: move radio item to models
