+++ yumako_jams/src/models/mod.rs
@@ -1,1 +1,3 @@
+pub mod radio_stream;
+/// Models of the radio configuration file
 pub mod radio_file;
