+++ yumako_jams/src/models/radio_file/error.rs
@@ -1,0 +1,11 @@
+#[derive(Debug, snafu::Snafu)]
+#[snafu(visibility(pub(super)))]
+pub enum RadioFileError {
+    #[snafu(display("Unknown step type `{step_type}`. Please check for typos"))]
+    UnknownStepTypeError {
+        step_type: String,
+
+        #[snafu(implicit)]
+        location: snafu::Location,
+    },
+}
