+++ yumako_jams/src/models/radio_file/mod.rs
@@ -1,3 +1,4 @@
+pub mod error;
 pub mod layer;
 pub mod radio;
 pub mod radio_input;
