+++ yumako_jams/src/modules/listen_data/last_listens.rs
@@ -16,7 +16,7 @@ use crate::client::YumakoClient;
 use crate::modules::listen_data::ListenAction;
 use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
-use crate::radio_item::RadioItem;
+use crate::models::radio_stream::radio_item::RadioItem;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct LatestListens {
