+++ yumako_jams/src/modules/scores/sort.rs
@@ -14,7 +14,7 @@ use crate::RadioStream;
 use crate::client::YumakoClient;
 use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
-use crate::radio_item::RadioItem;
+use crate::models::radio_stream::radio_item::RadioItem;
 use crate::radio_stream::RadioStreamaExt as _;
 
 #[derive(Serialize, Deserialize, Clone)]
