+++ yumako_jams/src/modules/seeders/listen_seeder.rs
@@ -9,7 +9,7 @@ use serde::Serialize;
 use crate::RadioStream;
 use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
-use crate::radio_item::RadioItem;
+use crate::models::radio_stream::radio_item::RadioItem;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct ListenSeeder {
