+++ yumako_jams/src/radio_stream.rs
@@ -12,8 +12,8 @@ use tuillez::pg_counted;
 use tuillez::pg_inc;
 use tuillez::tracing_indicatif::span_ext::IndicatifSpanExt;
 
+use crate::models::radio_stream::radio_item::RadioItem;
 use crate::modules::scores::ScoreMerging;
-use crate::radio_item::RadioItem;
 use crate::radio_variables::RadioVariables;
 
 /// The stream output of the radio
